% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-extract.R
\name{audit_recommendations}
\alias{audit_recommendations}
\title{Extract audit recommendations}
\usage{
audit_recommendations(audit)
}
\arguments{
\item{audit}{A \code{moat_audit} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{recommendation_id} and
\code{recommendation}.
}
\description{
Extract audit recommendations
}
\examples{
data("toy_moat")
audit <- moat(toy_moat, outcome = "outcome", batch = "batch", n_perm = 99)
audit_recommendations(audit)
}
