% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-object.R
\name{is_biome_audit}
\alias{is_biome_audit}
\title{Test if an object is a moat_audit}
\usage{
is_biome_audit(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if the object inherits from \code{moat_audit},
\code{FALSE} otherwise.
}
\description{
Test if an object is a moat_audit
}
\examples{
# This is an internal object structure, but can be tested:
audit <- moat:::biome_audit(risk = "low")
is_biome_audit(audit)
}
