% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moat-package.R
\docType{package}
\name{moat-package}
\alias{moat}
\alias{moat-package}
\title{moat: Microbiome Omics Audit Toolkit}
\description{
MOAT provides a pre-analysis audit layer for microbiome and omics studies:
before crossing into microbiome analysis, check the moat.
}
\examples{
# Load the toy dataset included in the package
data("toy_moat", package = "moat")

# Check its structure
class(toy_moat)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xec-cm/moat}
  \item \url{https://xec-cm.github.io/moat/}
  \item Report bugs at \url{https://github.com/xec-cm/moat/issues}
}

}
\author{
\strong{Maintainer}: Francesc Català-Moll \email{fcatala@irsicaixa.es} (\href{https://orcid.org/0000-0002-2354-8648}{ORCID})

}
\keyword{internal}
