% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-object.R
\name{print.moat_audit}
\alias{print.moat_audit}
\title{Print a moat_audit object}
\usage{
\method{print}{moat_audit}(x, ...)
}
\arguments{
\item{x}{A \code{moat_audit} object.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print a moat_audit object
}
\keyword{internal}
