% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-object.R
\name{print.safebiome_audit}
\alias{print.safebiome_audit}
\title{Print a safebiome_audit object}
\usage{
\method{print}{safebiome_audit}(x, ...)
}
\arguments{
\item{x}{A \code{safebiome_audit} object.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print a safebiome_audit object
}
\keyword{internal}
