% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebiome-package.R
\docType{package}
\name{safebiome-package}
\alias{safebiome}
\alias{safebiome-package}
\title{safebiome: Audit microbiome study designs before downstream analysis}
\description{
The package will provide tools to audit design confounding, batch effects,
correction feasibility, and validation leakage in microbiome studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xec-cm/safebiome}
  \item Report bugs at \url{https://github.com/xec-cm/safebiome/issues}
}

}
\author{
\strong{Maintainer}: Francesc Català-Moll \email{fcatala@irsicaixa.es} (\href{https://orcid.org/0000-0002-2354-8648}{ORCID})

}
\keyword{internal}
