% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-scoring.R
\name{summary.moat_audit}
\alias{summary.moat_audit}
\title{Print a MOAT audit summary}
\usage{
\method{summary}{moat_audit}(object, verbose = FALSE, ...)
}
\arguments{
\item{object}{A \code{moat_audit} object.}

\item{verbose}{A single logical value. When \code{TRUE}, include all module
reasons and recommendations. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{summary.moat_audit} object.
}
\description{
Print a MOAT audit summary
}
