% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-object.R
\name{validate_moat_audit}
\alias{validate_moat_audit}
\title{Validate a moat_audit object}
\usage{
validate_moat_audit(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
The validated \code{moat_audit} object.
}
\description{
Validate a moat_audit object
}
\keyword{internal}
