% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-object.R
\name{validate_biome_audit}
\alias{validate_biome_audit}
\title{Validate a safebiome_audit object}
\usage{
validate_biome_audit(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
The validated \code{safebiome_audit} object.
}
\description{
Validate a safebiome_audit object
}
\keyword{internal}
